Explain how fences interact with volatile operations
This was the result of a discussion with Amanieu at RustWeek.
It would be nice to add more details about the planned updates
(some ticket for example), and I'm not sure if this is the best
place to add it to the embedonomicon. But for now, I think
mentioning it here is better than nothing.