Initialize data and bss sections in assmebly
The embedonomicon currently suggests to initialize the bss and data
sections in Rust code. However, since this was written, there has been
extensive discussion about the soundness of the runtime initialization
code.

In particular, the following questions have been raised:
- Whether it is sound to write outside the bounds of the data pointed
  by pointers whose provenance come from a static variable.
- Whether it is sound to mutably alias all static memory and write to it
  during initialization.
- Whether it is unsound to enter the Rust abstract machine without the
  static variables being initialized.
- Whether pointers obtained from different static allocations (such as
  `_sbss` and `_ebss`) can be compared. Note that the code in
  embedonomicon does not suffer from this issue, since pointers are
  converted to `usize` values such that provenance does no longer apply
  by the time they are subtracted to determine the region size. This was
  not the case in multiple runtime crates.

The general consensus in the ecosystem has been to move away from
performing the initialization of the static memory regions in Rust.
In order to avoid creating a false sense of security while reading the
embedonomicon, these issues should be mentioned in the book, and the
code written in assembly to reflect the current best-practice approach
to platform initialization code.

References:
- https://github.com/rust-embedded/cortex-m-rt/issues/300
- https://github.com/rust-embedded/embedonomicon/issues/69
- https://rust-lang.zulipchat.com/#narrow/stream/136281-t-lang.2Fwg-unsafe-code-guidelines/topic/The.20least.20incorrect.20init.20code.20.3A)
- https://github.com/rust-lang/unsafe-code-guidelines/issues/259
- https://github.com/rust-embedded/cortex-m-rt/pull/301